requirements.txt
fastapi>=0.116.0
uvicorn>=0.35.0
sqlalchemy>=2.0.36
alembic>=1.14.0
jinja2>=3.1.4
python-multipart>=0.0.9
pydantic-settings>=2.6.1
pillow>=11.0.0
httpx>=0.28.0
